requirements.txt
mysql-connector-python == 9.2.0
flask==3.1.0
beautifulsoup4==4.13.3
